services/tenant-mgmt/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
sqlmodel>=0.0.21
asyncpg>=0.29.0
alembic>=1.13.0
aiokafka>=0.11.0
qdrant-client>=1.10.0
structlog>=24.1.0
pydantic-settings>=2.0.0
httpx>=0.27.0
